Continued work on speak.py and moved printing to console to own file
- Created console_log.py with 2 functions for printing to console
- Removed print_to_console from main.py
- Added better error handling to speak.py
-Fixes issue with speak.py not using different eleven_labs voices
