+++ requirements.txt
@@ -1,7 +1,7 @@
 beautifulsoup4
 colorama==0.4.6
 openai==0.27.2
-playsound==1.3.0
+playsound==1.2.2
 python-dotenv==1.0.0
 pyyaml==6.0
 readability-lxml==0.8.1
